requirements.txt
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
appnope==0.1.4
asttokens==3.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
comm==0.2.3
cryptography==46.0.7
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.18.0
duckduckgo_search==8.1.1
executing==2.2.1
fastjsonschema==2.21.2
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-genai==1.73.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
googlesearch-python==1.3.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.13
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
lazy-loader==0.5
libcst==1.8.6
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
moviepy==2.2.1
nbclient==0.10.4
nbconvert==7.17.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
oauthlib==3.3.1
openai==2.32.0
opencv-python-headless==4.13.0.92
opentelemetry-api==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
packaging==26.1
pandas==3.0.2
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pillow==11.3.0
pillow_heif==1.3.0
platformdirs==4.9.6
playwright==1.58.0
primp==1.2.3
proglog==0.1.12
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyee==13.0.1
Pygments==2.20.0
PyMuPDF==1.27.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
PyYAML-ft==8.0.0
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rpds-py==0.30.0
ruff==0.15.11
scikit-image==0.26.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
tabulate==0.10.0
tenacity==9.1.4
tifffile==2026.4.11
tinycss2==1.4.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
vision-agent==1.1.20
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
wikipedia==1.4.0
yt-dlp==2026.3.17
zipp==3.23.1
